refactor: SideBarMobile, SideBar에서 user 요청 한번으로 수정
