+++ frontend/src/components/common/BaseLayout/BaseLayout.js
@@ -9,23 +9,28 @@ import {
   NavBar,
   LogoContainer,
 } from './BaseLayout.styles';
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { LogoIcon, MenuIcon } from '../../../assets/icons';
-import { PATH, THEME_COLOR } from '../../../constants';
+import { LOCAL_STORAGE_KEY, PATH, SNACKBAR_MESSAGE, THEME_COLOR } from '../../../constants';
 import { useHistory, useLocation } from 'react-router-dom';
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import Button from '../Button/Button';
 import { BUTTON_BACKGROUND_TYPE } from '../Button/Button.styles';
 import { css } from '@emotion/react';
 import SideBarMobile from '../SideBarMobile/SideBarMobile';
+import { getMyInfoAsync, logout as logoutAction } from '../../../redux/authSlice';
+import { useSnackBar } from '../../../hooks';
 
 const BaseLayout = ({ children }) => {
   const history = useHistory();
   const location = useLocation();
+  const dispatch = useDispatch();
   const isLogin = !!useSelector((state) => state.authReducer.accessToken);
   const user = useSelector((state) => state.authReducer.user);
 
   const [isOpenSideBar, setIsOpenSideBar] = useState(false);
+
+  const { openSnackBar } = useSnackBar();
 
   const goMyPage = () => {
     history.push(PATH.MY_PAGE_ACCOUNT);
@@ -35,10 +40,31 @@ const BaseLayout = ({ children }) => {
     history.push(PATH.HOME);
   };
 
+  const logout = () => {
+    localStorage.removeItem(LOCAL_STORAGE_KEY.ACCESS_TOKEN);
+    dispatch(logoutAction());
+
+    openSnackBar(SNACKBAR_MESSAGE.SUCCESS_TO_LOGOUT);
+    setIsOpenSideBar(false);
+    history.push(PATH.HOME);
+  };
+
+  useEffect(() => {
+    const accessToken = JSON.parse(localStorage.getItem(LOCAL_STORAGE_KEY.ACCESS_TOKEN));
+
+    if (!accessToken) return;
+
+    dispatch(getMyInfoAsync(accessToken));
+  }, []);
+
   return (
     <Container>
-      <SideBar />
-      <SideBarMobile isOpenSideBar={isOpenSideBar} setIsOpenSideBar={setIsOpenSideBar} />
+      <SideBar logout={logout} />
+      <SideBarMobile
+        isOpenSideBar={isOpenSideBar}
+        setIsOpenSideBar={setIsOpenSideBar}
+        logout={logout}
+      />
       {/* mobile */}
       <NavBar isOpenSideBar={false}>
         <Button
